security: Remember owner only for top level image
Here is the problem: If all disks had XATTRs (i.e. domains using
them were started with owner remembering turned on) then
refcounting implemented in XATTRs would work nicely and we could
set the whole backing chain and restore it later. But world is
not that simple. As soon as there is one domain that was started
with the feature turned off (or simply by older libvirt), the
XATTR refounting does not reflect the actual number of uses by
running domains and therefore any attempt to restore might cut
off the old domain.

There is no simple way around this. Except artificially turning
the feature off for the rest of the backing chain.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <7b6246efd4c945fa40f67e31a1a0d07398050292.1553785388.git.mprivozn@redhat.com>