virerror: Make error reporting prettier
So far, if something goes wrong we print an error message, e.g.
like this:

  virsh # start fedora
  error: Failed to start domain fedora
  error: internal error: process exited while connecting to monitor: 2019-04-01T08:08:49.753850Z qemu-system-x86_64: -object memory-backend-memfd,id=ram-node0,hugetlb=yes,hugetlbsize=0,size=8589934592,host-nodes=0,policy=bind: Property 'memory-backend-memfd.hugetlbsize' doesn't take value '0'

This is very boring and usually too low level for users to know
what is going on or how to fix the problem. Let's reimplement our
error reporting then. After this patch the previous error turns
into (my favourite):

  virsh # start fedora
  error: Failed to start domain fedora
  error: operation failed: the printer thinks its a router.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <08725680ec84dc555ef0e360279673f872cee28e.1554106211.git.mprivozn@redhat.com>