virsh: Add JSON support to the 'define' subcommand
In order not to break user expectations (and scripts!) we need
to keep parsing input files as XML by default, so JSON support
is implemented as opt-in through the new --json option; a brand
new --xml option is also added for consistency.

That said, if the name of the input file ends in .json it's fair
to assume the contents are not going to be in XML format: we can
use this heuristic to make virsh more user-friendly by not
requiring the user to specify --json in this single case.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20190401123930.21777-8-abologna@redhat.com>