util: alloc: Clean also stack'd data with VIR_AUTOCLEAN
VIR_AUTOCLEAN uses __attribute__(cleanup) to clean stack'd data types.
The implementation calls the callback but did not clean the stack'd
structure besides to what the called function did. This adds a memset
call to clean it when unrolling the stack to achieve the same stack
cleanlines as with VIR_AUTOFREE and VIR_AUTOPTR.

The change from inline to ATTRIBUTE_UNUSED is necessary as gcc was
moaning that inlining would sometimes increase code size.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <e1f803f65fccb2677c27add3846898c59938e75d.1554123828.git.pkrempa@redhat.com>