util: Introduce 'virpin' module for dealing with PIN numbers
In https://www.redhat.com/archives/libvir-list/2019-February/msg01475.html
I was notified that I forgot to deal with the possibility that the JSON
number would contain jtomko's PIN number. I had to look around for a
leaked list of PIN numbers to avoid the problem described in the review.

This commit implements a checker which will allow deciding when to nuke
the JSON number string. We obviously have to spare a few CPU cycles to
make Jano feel safer.
Message-Id: <7139daeb647dcb57cf310d9194935bcf4b188b4b.1554123828.git.pkrempa@redhat.com>