setting QEMU_CAPS_WAKEUP_SUSPEND_SUPPORT qemu capability
QEMU_CAPS_WAKEUP_SUSPEND_SUPPORT is a bool that represents the value
of the wakeup-suspend-support flag of the new query-current-machine
QMP API, available in QEMU 4.0.

The flag is set in qemuProcessLaunch, using a new function called
qemuProcessSetupWakeupSuspendSupport. This function checks first
if the QEMU process has support for the new API, which is reflected
by the QEMU_CAPS_QUERY_CURRENT_MACHINE cap. If positive, query it
and set QEMU_CAPS_WAKEUP_SUSPEND_SUPPORT with the current value
of wakeup-suspend-support.
Message-Id: <20190401211826.8598-3-danielhb413@gmail.com>