docs: Remove search.php and all references
libvirt.org/search.php drops into some kind of screen which I guess
is supposed to show a search bar with options, but presently for me
renders as nothing but the following text:

Search the documentation on Libvirt.org

The search service indexes the libvirt APIs and documentation as well as the libvir-list@redhat.com mailing-list archives. To use it simply provide a set of keywords:

The main page search bar now redirects to google, this page is broken,
I say we just remove it and move on.

Signed-off-by: Cole Robinson <crobinso@redhat.com>
Message-Id: <a16bde9c587dd1d5813d0b855e67d5c7410ca5c5.1554329992.git.crobinso@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>