qemu_process: setting QEMU_CAPS_QUERY_CURRENT_MACHINE caps
The new QMP API query-current-machine reports VM capabilities that
can vary depending on the VM configuration, in contrast to most
QEMU caps that can be set once during libvirtd init as exposed in
qemu_capabilities.c.

As such, to properly set the runtime capabilities exposed by this
API, we must query them at VM launch time. To do that, a new function
called qemuProcessSetCurrentMachineCaps is called in qemuProcessLaunch.
A new struct called qemuMonitorCurrentMachineInfo was created to host
the output of the API (which contains only the wakeup-suspend-support
flag for now, but can be expanded in the future). This struct is
populated by qemuMonitorGetCurrentMachineInfo, which does the heavy
work - executes query-current-machine, parses the result and populates
the qemuMonitorCurrentMachineInfo struct.

qemuProcessSetCurrentMachineCaps can then read the structure and
set the capabilities accordingly.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20190409161816.21542-3-danielhb413@gmail.com>