qemuDomainPMSuspendForDuration: check for QEMU_CAPS_PM_WAKEUP_SUPPORT
If the current QEMU guest can't wake up from suspend properly,
avoid suspending the guest at all. This is done by checking the
QEMU_CAPS_PM_WAKEUP_SUPPORT cap.

The absence of the cap indicates that we're dealing with a QEMU
version older than 4.0 (which implements the required QMP API).
In this case, proceed as usual with the suspend logic since
we can't assume whether the guest has support or not.

This is the output of dompmsuspend in a guest that does not
have wake-up support declared in the query-current-machine:

$ sudo ./run tools/virsh dompmsuspend ub1810-noACPI3 mem
error: Domain ub1810-noACPI3 could not be suspended
error: this function is not supported by the connection driver: Domain does not have suspend support

Fixes: https://bugs.launchpad.net/ubuntu/+source/qemu/+bug/1759509
Reported-by: Balamuruhan S <bala24@linux.vnet.ibm.com>
Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20190409161816.21542-4-danielhb413@gmail.com>