domain_conf: allow CDROM/Floppy media change with attach-device
Commit f1d6585300 ("domain_conf: check device address before attach")
added an address verification for all devices, avoiding calling
the driver attach() for a case in which we know that the device
would have duplicated address.

That commit failed to contemplate the case of CDROM/Floppy devices
that can change and eject media using attach-device, an usage
that is also covered by this command. This patch adds an
extra condition for the added address verification code,
allowing CDROM/Floppy devices to bypass it.

Fixes: f1d6585300001c7b23b8796a0faa4411c3531996
Reported-by: Jim Fehlig <jfehlig@suse.com>
Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20190410214907.27170-3-danielhb413@gmail.com>