qemu: monitor: implement query-cpu-model-comparison
Interfaces with QEMU to compare CPU models. The command takes two
CPU models, A and B, that are given a model name and an optional list
of CPU features. Through the query-cpu-model-comparison command issued
via QMP, a result is produced that contains the comparison evaluation
string (identical, superset, subset [s390x specific], incompatible)
as well as a list of properties (aka CPU features) responsible for
the subset or superset result.

Though the "responsible properties" are captured, they are not
reported by libvirt.

This command is hooked into the virsh hypervisor-cpu-compare command.
As such, the CPU model XML provided to the command will be compared
to the hypervisor CPU contained in the QEMU capabilities file for the
appropriate QEMU binary (for s390x, this CPU definition can be observed
via virsh domcapabilities).

s390x can report that the first model (A) is a subset of the second
(B). If A is the hypervisor CPU and a subset of B (the CPU contained
in the XML), then B would not be able to run on this machine and thus
we will report that CPU model B is incompatible.

Signed-off-by: Collin Walling <walling@linux.ibm.com>
Message-Id: <20190415205925.28737-5-walling@linux.ibm.com>
Tested-by: Daniel Henrique Barboza <danielhb413@gmail.com>