backup: Introduce virDomainBackup APIs
Introduce a few new public APIs related to incremental backups.  This
builds on the previous notion of a checkpoint (without an existing
checkpoint, the new API is a full backup, differing from
virDomainBlockCopy in the point of time chosen and in operation on
multiple disks at once); and also allows creation of a new checkpoint
at the same time as starting the backup (after all, an incremental
backup is only useful if it covers the state since the previous
backup).  Snapshot creation is also a point in time at which creating
a checkpoint atomically can be useful; as checkpoints are independent
objects, it is not worth embedding <domaincheckpoint> inside
<domainsnapshot>, and therefore we need a more powerful version of
virDomainSnapshotCreateXML(), where we borrow from the naming pattern
of virDomainMigrate2() and friends.

A backup job also affects filtering a listing of domains, as well as
adding event reporting for signaling when a push model backup
completes (where the hypervisor creates the backup); note that the
pull model does not have an event (starting the backup lets a third
party access the data, and only the third party knows when it is
finished).

Since multiple backup jobs can be run in parallel in the future (well,
qemu doesn't support it yet, but we don't want to preclude the idea),
virDomainBackupBegin() returns a positive job id, and the id is also
visible in the backup XML. But until a future libvirt release adds a
bunch of APIs related to parallel job management where job ids will
actually matter, the documentation is also clear that job id 0 means
the 'currently running backup job' (provided one exists), for use in
virDomainBackupGetXMLDesc() and virDomainBackupEnd().

The full list of new APIs:
        virDomainBackupBegin;
        virDomainBackupEnd;
        virDomainBackupGetXMLDesc;
        virDomainSnapshotCreateXML2;

Signed-off-by: Eric Blake <eblake@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20190417140921.11964-5-eblake@redhat.com>