backup: Allow for lists of checkpoint objects
Create a new file for managing a list of checkpoint objects, borrowing
heavily from existing virDomainSnapshotObjList paradigms.

Note that while checkpoints definitely have a use case for multiple
children to a single parent (create a base snapshot, create a child
snapshot, revert to the base, then create another child snapshot),
it's harder to predict how checkpoints will play out with reverting to
prior points in time. Thus, in initial use, we may find that in a list
of checkpoints, you never have more than one child.  However, as the
snapshot machinery is already generic, it is easier to reuse the
generic machinery that tracks relations between domain moments than it
is to open-code a new list-management scheme just for checkpoints.

The virDomainMomentObjList code is not quite polymorphic enough until
we patch virDomainMomentDef to be a proper virObject, but doing that
is a bigger audit.  So for now, I had to duplicate the cleanup calls
based on a bool flag for which type needs cleaning. Oh well.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190417140921.11964-8-eblake@redhat.com>