backup: Implement backup APIs for remote driver
The remote code generator had to be taught about the new
virDomainCheckpointPtr type, and about the capitalization of
virDomainSnapshotCreateXML2(), at which point the remote driver code
for backups can be generated.

Note that virDomainSnapshotCreateXML2 has an unconditional ACL check
on domain:checkpoint privileges, even though those privileges are only
needed when the checkpointXml argument is non-NULL. If this turns out
to be a problem in practice, we will have to either enhance the access
code generator to allow conditionals on whether a parameter is
non-NULL, or else create a witness flag (similar to
VIR_TYPED_PARAM_STRING_OKAY) that gets automatically set by
libvirt-domain-snapshot.c when checkpointXml is non-NULL (so that we
can then reuse the existing syntax for making checks conditional on
particular bits in flags). But until someone complains, being
theoretically too strict is good enough.

Signed-off-by: Eric Blake <eblake@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20190417140921.11964-10-eblake@redhat.com>